requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==5.0.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
djangorestframework==3.14.0
pillow==10.2.0
pytz==2023.3.post1
sqlparse==0.4.4
